+++ src/spyglass/spikesorting/spikesorting_merge.py
@@ -83,6 +83,8 @@ class SpikeSortingOutput(_Merge, SpyglassMixin):
         merge_ids : list
             list of merge ids from the restricted sources
         """
+        # TODO: replace with long-distance restrictions
+
         merge_ids = []
 
         if "v1" in sources:
@@ -185,6 +187,7 @@ class SpikeSortingOutput(_Merge, SpyglassMixin):
         return part_table * sort_group_info  # join the info with merge id's
 
     def get_spike_times(self, key):
+        """Get spike times for the group"""
         spike_times = []
         for nwb_file in self.fetch_nwb(key):
             # V1 uses 'object_id', V0 uses 'units'