+++ tests/utils/test_mixin.py
@@ -39,6 +39,11 @@ def test_auto_increment(schema_test, Mixin):
     assert ret["id"] == 2, "Auto increment not working."
 
 
+def test_good_file_like(common):
+    common.Session().file_like("min")
+    assert len(common.Session()) > 0, "file_like not working."
+
+
 def test_null_file_like(schema_test, Mixin):
     schema_test(Mixin)
     ret = Mixin().file_like(None)
@@ -52,60 +57,11 @@ def test_bad_file_like(caplog, schema_test, Mixin):
     assert "No file_like field" in caplog.text, "No warning issued."
 
 
-def test_partmaster_detect(Nwbfile, pos_merge_tables):
-    """Test that the mixin can detect merge children of merge."""
-    assert len(Nwbfile._part_masters) >= 14, "Part masters not detected."
-
-
-def test_downstream_restrict(
-    Nwbfile, frequent_imports, pos_merge_tables, lin_v1, lfp_merge_key
-):
-    """Test that the mixin can join merge chains."""
-
-    _ = frequent_imports  # graph for cascade
-    _ = lin_v1, lfp_merge_key  # merge tables populated
-
-    restr_ddp = Nwbfile.ddp(dry_run=True, reload_cache=True)
-    end_len = [len(ft) for ft in restr_ddp]
-
-    assert sum(end_len) >= 8, "Downstream parts not restricted correctly."
-
-
-def test_get_downstream_merge(Nwbfile, pos_merge_tables):
-    """Test that the mixin can get the chain of a merge."""
-    lin_output = pos_merge_tables[1].full_table_name
-    assert lin_output in Nwbfile._part_masters, "Merge not found."
-
-
 @pytest.mark.skipif(not VERBOSE, reason="No logging to test when quiet-spy.")
-def test_ddp_warning(Nwbfile, caplog):
-    """Test that the mixin warns on empty delete_downstream_merge."""
-    (Nwbfile.file_like("BadName")).delete_downstream_parts(
-        reload_cache=True, disable_warnings=False
-    )
-    assert "No part deletes found" in caplog.text, "No warning issued."
-
-
-def test_ddp_dry_run(
-    Nwbfile, frequent_imports, common, sgp, pos_merge_tables, lin_v1
-):
-    """Test that the mixin can dry run delete_downstream_merge."""
-    _ = lin_v1  # merge tables populated
-    _ = frequent_imports  # graph for cascade
-
-    pos_output_name = pos_merge_tables[0].full_table_name
-
-    param_field = "trodes_pos_params_name"
-    trodes_params = sgp.v1.TrodesPosParams()
-
-    rft = [
-        table
-        for table in (trodes_params & f'{param_field} LIKE "%ups%"').ddp(
-            reload_cache=True, dry_run=True
-        )
-        if table.full_table_name == pos_output_name
-    ]
-    assert len(rft) == 1, "ddp did not return restricted table."
+def test_insert_fail(caplog, common, mini_dict):
+    this_key = dict(mini_dict, interval_list_name="BadName")
+    common.PositionSource().find_insert_fail(this_key)
+    assert "IntervalList: MISSING" in caplog.text, "No warning issued."
 
 
 def test_exp_summary(Nwbfile):
@@ -114,11 +70,71 @@ def test_exp_summary(Nwbfile):
     assert fields == expected, "Exp summary fields not as expected."
 
 
+def test_exp_summary_no_link(schema_test, Mixin):
+    schema_test(Mixin)
+    assert Mixin()._get_exp_summary() is None, "Exp summary not None."
+
+
+def test_exp_summary_auto_link(common):
+    lab_member = common.LabMember()
+    summary_names = lab_member._get_exp_summary().heading.names
+    join_names = (lab_member * common.Session.Experimenter).heading.names
+    assert summary_names == join_names, "Auto link not working."
+
+
 def test_cautious_del_dry_run(Nwbfile, frequent_imports):
     _ = frequent_imports  # part of cascade, need import
-    ret = Nwbfile.cdel(dry_run=True)
-    part_master_names = [t.full_table_name for t in ret[0]]
-    part_masters = Nwbfile._part_masters
-    assert all(
-        [pm in part_masters for pm in part_master_names]
-    ), "Non part masters found in cautious delete dry run."
+    ret = Nwbfile.cautious_delete(dry_run=True)[1].full_table_name
+    assert (
+        ret == "`common_nwbfile`.`~external_raw`"
+    ), "Dry run delete not working."
+
+
+@pytest.mark.skipif(not VERBOSE, reason="No logging to test when quiet-spy.")
+def test_empty_cautious_del(caplog, schema_test, Mixin):
+    schema_test(Mixin)
+    Mixin().cautious_delete(safemode=False)
+    Mixin().cautious_delete(safemode=False)
+    assert "empty" in caplog.text, "No warning issued."
+
+
+def test_super_delete(schema_test, Mixin, common):
+    schema_test(Mixin)
+    Mixin().insert1((0,), skip_duplicates=True)
+    Mixin().super_delete(safemode=False)
+    assert len(Mixin()) == 0, "Super delete not working."
+
+    logged_dels = common.common_usage.CautiousDelete & 'restriction LIKE "Sup%"'
+    assert len(logged_dels) > 0, "Super delete not logged."
+
+
+def test_compare_versions(common):
+    # Does nothing in test_mode
+    compare_func = common.Nwbfile().compare_versions
+    compare_func("0.1.0", "0.1.1")
+
+
+@pytest.fixture
+def custom_table():
+    """Custom table on user prefix for testing load_shared_schemas."""
+    db, table = dj.config["database.user"] + "_test", "custom"
+    dj.conn().query(f"CREATE DATABASE IF NOT EXISTS {db};")
+    dj.conn().query(f"USE {db};")
+    dj.conn().query(
+        f"CREATE TABLE IF NOT EXISTS {table} ( "
+        + "`merge_id` binary(16) NOT NULL COMMENT ':uuid:', "
+        + "`unit_id` int NOT NULL, "
+        + "PRIMARY KEY (`merge_id`), "
+        + "CONSTRAINT `unit_annotation_ibfk_1` FOREIGN KEY (`merge_id`)  "
+        + "REFERENCES `spikesorting_merge`.`spike_sorting_output` (`merge_id`) "
+        + "ON DELETE RESTRICT ON UPDATE CASCADE);"
+    )
+    yield f"`{db}`.`{table}`"
+
+
+def test_load_shared_schemas(common, custom_table):
+    # from spyglass.common import Nwbfile
+
+    common.Nwbfile().load_shared_schemas(additional_prefixes=["test"])
+    nodes = common.Nwbfile().connection.dependencies.nodes
+    assert custom_table in nodes, "Custom table not loaded."